[Nailang] Add string formating. [Nailang] Add variable existance detection. [NLCL] Add logging support (printf). [test] Force NailangTest to use C++17 since gtest 1.10.x is not compatible with C++20.

+++ Tests/NailangTest/ParserContextTest.cpp
@@ -243,7 +243,7 @@ TEST(ParserCtx, ParseFunc)
                     args.emplace_back(arg);
                     continue;
                 case special::CharEnd:
-                    EXPECT_NE(next, special::CharEnd);
+                    EXPECT_NE((uint32_t)next, (uint32_t)special::CharEnd);
                     return { };
                 default:
                     EXPECT_TRUE(false);
